change error trace to info in unit state check
The getUnitState() function was made more generic recently to handle
looking at other units. In some cases, a unit like the
obmc-bmc-service-quiesce@.target will not be loaded so systemd will not
be able to return any information to the GetUnit call.

This should not be treated like an error so change the log to an info
one. Callers of the function properly handle the empty string return.

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>
Change-Id: Iadbb95acc1bce2b383f1339c9bf641c12c93787b