RBMC: Add serialization utilities
Add data::read() and data::write() to use for reading and writing
key/value pairs in
/var/lib/phosphor-state-manager/redundant-bmc/data.json.

A JSON file was chosen mostly just so it's easy to see all the values at
once and to capture them all in a dump or error log.

To start with, these will be used to save/restore the previous role and
if there was a problem that forced the BMC to be passive.

Tested:
New unit tests pass.

Change-Id: I66dfbf3cb82f454aade9347d7d4b7f2ac3b3e5f9
Signed-off-by: Matt Spinler <spinler@us.ibm.com>