RBMC: Start obmc-bmc-active.target if active
After the BMC has determined it has the active role, start the new
obmc-bmc-active.target systemd target.  In the future, systemd services
that only run when the BMC is active will be linked to it and so will
now start.

This class also introduces ActiveRoleHandler and PassiveRoleHandler
classes that will house all of the role specific code so that the
Manager class doesn't have to handle everything.

The starting of the target will be in the ActiveRoleHandler class as the
first thing it does when it starts.

It waits for the target to finish (or fail) starting because after this
is enabling redundancy which needs to wait for all of the active
services to be running.

It doesn't use a PropertiesChanged or JobRemoved signal from systemd to
wait for the target to be done because those have to be subscribed to
first, and OpenBMC has problems with that Subscribe method call taking
more than 25 seconds during startup which doesn't seem like a good trade
off.

Tested:

```
phosphor-rbmc-state-manager[938]: Starting unit obmc-bmc-active.target
phosphor-rbmc-state-manager[938]: Finished waiting for obmc-bmc-active.target to start (result = active)
```

Signed-off-by: Matt Spinler <spinler@us.ibm.com>