RBMC: Start obmc-bmc-passive.target if passive
Just like obmc-bmc-active.target is started for the active BMC, start
obmc-bmc-passive.target if the BMC is determined to have the passive
role.  Services that only need to run when the BMC is passive can key
off of this target.

The two targets conflict with each other, so if the active target is
started when the passive one is already running, the passive one will
stop.

Tested:
```
phosphor-rbmc-state-manager[636]: Starting unit obmc-bmc-passive.target
phosphor-rbmc-state-manager[636]: Finished waiting for obmc-bmc-passive.target to start (result = active)
```

```
$ systemctl status obmc-bmc-passive.target
* obmc-bmc-passive.target - Passive BMC Target
     Loaded: loaded (/usr/lib/systemd/system/obmc-bmc-passive.target; static)
     Active: active since Mon 2024-09-30 20:04:43 UTC; 50s ago
```

Signed-off-by: Matt Spinler <spinler@us.ibm.com>