RBMC: Support for getting chassis power state
The power state of the system will be needed in a few different places.
For example the property used to manually disable redundancy cannot be
modified when the system is powered on.  Also, fail overs will be
disabled in the window between power on and boot complete.

Add a 'isPoweredOn()' function to the Services class to return this
information.  This will be obtained from the CurrentPowerState property
on the chassis0 state object, which represents the overall power state
of the system.

This power value will be cached, so there are InterfacesAdded and
PropertiesChanged handlers for it.  In the future there may also need to
be code that runs when the power state changes, and the
PropertiesChanged handler provides a place for that.

It was verified that the chassis state manager does emit an
InterfacesAdded signal on startup.

The chassis management code will not be started until the
obmc-bmc-active target is started, so there will be no power state value
available before then and the function will throw an exception until the
power state has been obtained from D-Bus.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>