RMBC: Create HealthMonitor class
This class monitors a boolean health status and will execute callbacks
when the health is bad after configured amounts of time, and when it
recovers.  It will be used for monitoring the passive BMC's heartbeat,
so that actions can be taken when that BMC dies and recovers.  It may
also be used for monitoring the passive BMC's network health.

It is told about health status change by having its setHealthStatus
method called.

It will then call the callback passed into the constructor at these
points:
- Health warning
- Health critical
- Health recovered

It provides startMonitor() and stopMonitor() methods to start and stop
monitoring.

More details are in the header file.

Tested:
New unit tests pass.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>