RBMC: Add rbmctool option to reset sibling BMC
Add an option to rbmctool to reset the sibling BMC using the reset
GPIO.  This reset will also be done during a failover, and the class
created to do the reset will also be used in that case.

Note that in the assertReset() and releaseReset() functions the GPIO is
not held, so that during real use during a failover if something goes
horribly wrong the GPIO won't get stuck being held.

Just a long option was provided to attempt to make it a bit harder to
type by accident.

Tested:

New options:

Options:
  -h,--help                   Print this help message and exit
  -d                          Display RBMC Information
  -e Needs: -d                Add extended RBMC details to the display
  --reset-sibling             Reset the sibling BMC

The reset:

$ ./rbmctool --reset-sibling
<6> Asserting sibling BMC reset GPIO
<6> Releasing sibling BMC reset GPIO

Change-Id: I44e61de36e8e83cbba93913ac643261b137345bc
Signed-off-by: Matt Spinler <spinler@us.ibm.com>