RBMC: Create Providers class to hold helpers
Create a Providers class to hold the various system access classes like
Sibling and Services to simplify creating them and passing them around.
For now there are just two of these classes, but there are more coming
in the future (such as for resetting the sibling and managing data
sync).

The Providers class itself is abstract, and providers access to the base
class versions of the helpers.  ProvidersImpl is the implementation
which will provide access to the real SiblingImpl and ServicesImpl
classes.  If there is ever a desire to mock all of these classes a
ProvidersMock could just return the mock versions of those classes.

Tested:
Everything still works.

Change-Id: I05dd9388b653e6a40b711542232ba605ed58f841
Signed-off-by: Matt Spinler <spinler@us.ibm.com>