RBMC: Remove sibling pos check from role det
The hardware is wired as such that there is no longer a concern about
both BMCs reporting the same position, so remove the code that was
checking for it.

A bigger concern would be a BMC failing to obtain its position, in which
case it should just make itself be passive. That will be dealt with in
the future.

Change-Id: I3b634678a84d0a7bf05d90c312cc92844e7ea0c3
Signed-off-by: Matt Spinler <spinler@us.ibm.com>