RBMC: Remove sibling's 'sibling comm OK' check
It turns out that when one BMC can't talk to another one when it comes
up, it will wait until it hits its timeouts and then just become active
because it thinks its the only working BMC.

For that reason, there is no need to check if the other BMC can talk to
the current BMC when enabling redundancy, as it would never get to that
point anyway.

This does lead to a case where both BMCs can think they are active which
still needs to to be solved.

Change-Id: Ied5fede22f6adee4ebb1402ab11588f4eb9e9f2f
Signed-off-by: Matt Spinler <spinler@us.ibm.com>