RBMC: Check if StartFailover method can proceed
When the StartFailover method is called, add the following checks and
return the 'Unavailable' D-Bus error if any fail:

1. Check the RoleHandler object has been created
2. Check that a failover isn't already in progress
3. Check if the role handler has any reasons to block failovers.

On the passive BMC, check 3 will call the new getFailoverBlockedReason()
function that will check if failovers are allowed, etc.  On the active
BMC, it will just fail stating the reason is the BMC isn't passive.  In
the future this would change when we need to allow failovers from the
active BMC.

Tested:

Good path that falls through to last check:
```
phosphor-rbmc-state-manager[1835]: Failovers are not implemented yet
```

Calling on the active BMC:
```
phosphor-rbmc-state-manager[2387]: Active BMC cannot trigger a failover now
phosphor-rbmc-state-manager[2387]: Failover is blocked because: This BMC is not passive
```

FailoversAllowed is false:
```
phosphor-rbmc-state-manager[1835]: Failover is blocked because: Failovers are not allowed
```

Forcing a failover when FailoversAllowed is false.  This gets through:
```
phosphor-rbmc-state-manager[1835]: The failover 'Force' option is set while failovers are not allowed
phosphor-rbmc-state-manager[1835]: Failovers are not implemented yet
```

Active BMC not running, redundancy was last known to be enabled:
```
phosphor-rbmc-state-manager[1835]: There is no sibling heartbeat but redundancy was last known to be enabled
phosphor-rbmc-state-manager[1835]: Failovers are not implemented yet
```

Active BMC not running, redundancy was last known to be disabled:
```
phosphor-rbmc-state-manager[1924]: Failover is blocked because: Sibling is dead but redundancy wasn't previously enabled
```

Change-Id: I56a413981c426d74b0ce4476057aa73dc1c33284
Signed-off-by: Matt Spinler <spinler@us.ibm.com>