RBMC: Add toggleReset function to SiblingReset
Put the calls to set the reset GPIO both to a 1 and a 0 into the same
function, toggleReset, so the GPIO only has to be requested and released
once.

The standalone assert and release reset functions are being kept for use
in BMC card concurrent maintenance functionality.

Also add the SiblingReset class to the Providers class so code will have
access to it.  It will be used during a failover to reset the original
passive BMC so it can reboot and become passive.

Tested:
rbmctool --reset-sibling still works:

```
$ rbmctool --reset-sibling
<6> Toggling sibling reset GPIO

// journal:
rbmctool[1262]: Toggling sibling reset GPIO
phosphor-rbmc-state-manager[1143]: Sibling property Active changed
phosphor-rbmc-state-manager[1143]: Sibling BMC heartbeat lost
```

Change-Id: I0c15712c1720a06e6f72a08bdbfd8e3e937b4eff
Signed-off-by: Matt Spinler <spinler@us.ibm.com>