RBMC: Add FO-in-progress as no redundancy reason
In a failover, the passive BMC will reset the original active BMC and
then claim the active role itself and wait for the other BMC to come
back as passive.

During that time, redundancy won't be enabled as the sibling BMC is in
the middle of rebooting.  A new no redundancy reason of 'failover in
progress' is being added that will be used during this time to make it
obvious what is going on.

In addition, a new function 'disableRedundancyDueToFailover' is being
added to RedundancyMgr that can be called to disable redundancy with
this reason.  The reason would then persist on future calls to determine
redundancy until the other new function, clearFailoverInProgress, is
called.

Tested:
The new unit test that checks this reason passes.

Change-Id: Ia481ad1a74616625735c2a31f8efc737c58b1681
Signed-off-by: Matt Spinler <spinler@us.ibm.com>