RBMC: Add service to wait for role target
Add a new script wait-for-active-passive-target.sh that will wait for
the role to be determined on D-Bus and then wait for either the
obmc-bmc-active.target or obmc-bmc-passive.target to be started.

Also add a new service phosphor-wait-for-active-passive-target.service
to call it.

This service will be linked to multi-user.target, so it can keep it from
completing until the script is done.  This way, the BMC ready state can
be held off until that is done.  This avoids the case where the BMC
state says ready but it actually isn't ready because either the role is
still unknown or all of the active services haven't been started yet.

Tested:

Waits for almost 2 minutes:
```
20:07:38 p11bmc systemd[1]: Starting Wait for obmc-bmc-active or obmc-bmc-passive target to be started
20:09:33 p11bmc wait-for-active-passive-target.sh[858]: BMC role determined, now waiting for obmc-bmc-passive.target to start
20:09:33 p11bmc wait-for-active-passive-target.sh[858]: Done waiting for obmc-bmc-passive.target
20:09:33 p11bmc systemd[1]: phosphor-wait-for-active-passive-target.service: Deactivated successfully.
20:09:33 p11bmc systemd[1]: Finished Wait for obmc-bmc-active or obmc-bmc-passive target to be started.
```

Change-Id: I4804b9a375b6267fca8a9f3f80a19bb44fbc313f
Signed-off-by: Matt Spinler <spinler@us.ibm.com>