RBMC: Add FO-in-progress check for FO not allowed
In a failover, the passive BMC will reset the original active BMC and
then claim the active role itself and wait for the other BMC to come
back as passive.

During that time, the FailoversAllowed D-Bus property will be set to
false to make it obvious to users that a failover isn't available then.
When the passive BMC makes it back to ready state and the active BMC
checks if redundancy can still be enabled, FailoversAllowed can be set
back to true assuming nothing else is blocking it.

To handle this, this commit adds support for a 'failover-in-progress'
reason for failovers to be disallowed.

It also adds a clearFailoversAllowedDuringFailover function on
RedundancyMgr that will be called during the failover to call
determineAndSetFailoversAllowed with failoverInProgress = true.

Later, RedundancyMgr::clearFailoverInProgress can be called to clear the
failover in progress indication.

Tested:
New testcases pass.

Change-Id: I1107d1a67dba1c7ea003758d95055d2affcae4e7
Signed-off-by: Matt Spinler <spinler@us.ibm.com>