RBMC: Log failover requesters/timestamps
The 'requester' parameter of the StartFailover method says who is
requesting the failover.  Save that and the timestamp the failover was
requested in a file so that it can be displayed by rbmctool and also
captured in BMC dumps.

A new 'logFailover' function is used for this. It is called by the BMC
driving the failover after the point of no return which is after the
original active BMC is reset.  It writes to a JSON file that keeps the
most recent 10 failovers.

The file is in the same directory the existing 'data.json' file is in,
and is named bmcN_failovers, where N is the BMC position.  To get a full
view of every failover in the system from a single BMC, the files would
need to be synced so each BMC has the file for both positions.

The 'rbmctool -de' option will now display the requester and timestamp
of the most recent failover when run on the active BMC. It doesn't do
this on the passive BMC because the most recent failover saved on that
BMC isn't the actual most recent one, as that would be on the current
active BMC.  If the files are synced in the future, then possibly it
could use the file synced over from the active BMC.

Tested:

Most recent failover is shown:
```
Local BMC
-----------------------------
Role:                 Active
BMC Position:         1
Redundancy Enabled:   true
BMC State:            Ready
Failovers Allowed:    true
Failover In Progress: false
FW Version Hash:      7D719923
Provisioned:          true
Role Reason:          Failover
Last failover driven by this BMC:
    Requester: Tool                    <----
    Timestamp: 2025-10-08 17:14:24 UTC <----

```

And the file itself:
```
$ cat /var/lib/phosphor-state-manager/redundant-bmc/bmc1_failovers
[
    {
        "requester": "Tool",
        "timestamp": "2025-10-08 16:23:54 UTC"
    },
    {
        "requester": "Tool",
        "timestamp": "2025-10-08 17:14:24 UTC"
    }
]
```

Also, new testcases pass.

Change-Id: I1fc538c0f5fcb3e14f7b1159d23f19717c350820
Signed-off-by: Matt Spinler <spinler@us.ibm.com>