RBMC: Use Available prop to get sib BMC presence
Previously, the Sibling::isBMCPresent check was just hardcoded to return
true, which of course doesn't work if the BMC may not be present.

Ideally there would be GPIOs that could be used for this, however not
all of IBM's redundant BMC systems have them, and the one that will
isn't far enough along to code against.

So for now, just look at the Availability interface on the state/bmc1
object path which reflects if the communication path to the sibling BMC
was detected.

When more information is known about the presence GPIO, it can be
checked first and then code can fall back to the availability check if
the GPIO read fails due to a hardware error or if the system doesn't
have a GPIO.

Add this new interface and property to the Sibling handling of the other
interfaces on that object path which already watch for all the D-Bus
signals to track changes, and return the value of Available in the
isBMCPresent() function.

The interface will be present as soon as the sibling daemon starts up
regardless of if the BMC is responding or not.

Tested:

When available, works the same as before.

When not available:
```
Local BMC
-----------------------------
Role:                 Active
BMC Position:         0
Redundancy Enabled:   false
BMC State:            Ready
Failovers Allowed:    false
Failover In Progress: false
FW Version Hash:      76A7D135
Provisioned:          true
Role Reason:          Sibling not alive
Reasons for no BMC redundancy:
    Sibling is missing

Sibling BMC
-----------------------------
Sibling data not available
```

Change-Id: I5bdfa18d8e15b261c75e9f55a605d58cc115617c
Signed-off-by: Matt Spinler <spinler@us.ibm.com>