Revert "phosphor-discover-system-state: set after multi-user.target"
This reverts commit 1901ab743aa3f820a304299aa4aea6c2827ba1de.

This trick to have phosphor-discover-system-state@.service run after
multi-user.target does not work well within the redundant BMC design.
End up in a situation where
phosphor-wait-for-active-passive-target.service starts and is waiting on
obmc-bmc-active.target which is waiting on
phosphor-discover-system-state.service which is waiting on
multi-user.target. Everything then just hangs up on these
services/targets.

We already have a delay and an option to tell discover-system-state to
wait for BMC Ready so this change was not something IBM systems needed.

Remove for now and figure out how to upstream this at a later date.