RBMC: Improvements to red override prop handling
The first improvement is to print a unique trace when someone tries to
set the disable redundancy override property but the code isn't far
enough along to know the system state.  Before, that path would print a
trace that it cannot be modified when power is on.

The second improvement is to still let the code flow to the
determineRedundancyAndSync() function when someone is trying to manually
disable redundancy but redundancy is already disabled.

Previously the code would just return early, but that would prevent the
manual override being added as a redundancy disabled reason.

Change-Id: Id1a23c080efeea15287be09b6327e9fdb0523099
Signed-off-by: Matt Spinler <spinler@us.ibm.com>