RBMC: Move FOInProgress persist to iface class
The FailoverInProgress D-Bus property needs to be persisted every time
it is changed, so just move that code into the property set handler for
that property in the RedundancyInterface class.  Then it is always just
done in one place and can be removed from the others.

This requires a small change to the failover flow to not set
FailoverInProgress = true until after the original active BMC has been
reset, since we don't want to persist it until then.  That reset is fast
though so nobody should be able to notice a difference.

Tested:

Property is still persisted:

```
{
    "FailoverInProgress": true,  <-----------
    "FailoversNotAllowedReasons": [
        "A failover is in progress"
    ],
    "PassiveDueToError": false,
    "Role": 1,
    "RoleReason": "Failover"
}
```

Reboots during a failover still work correctly:

on active:
```
phosphor-rbmc-state-manager[843]: Failover was previously in progress
```

and on passive

```
Role Reason:          Sibling was driving a failover
```

Change-Id: Ie63b3a5eff5d0011d5c13de123fa30b03605841b
Signed-off-by: Matt Spinler <spinler@us.ibm.com>