RBMC: Add API for sibling 'reason for no red'
Now that the sibling daemon fills in the 'ReasonsForNoRedundancy'
property on the state/bmc1 object path to indicate that the passive BMC
has a reason it should not be active so redundancy cannot be enabled,
add an API for that to the Sibling class so it can be used by other
code.

The sibling daemon just adds a single entry to the std::set property
when it gets the indication from the sibling BMC that it can't be
active.

The new sibling API is a 'bool getHasReasonForNoRedundancy()' function
that will return true when that property isn't empty.  This will be used
by upcoming code as an additional check when enabling redundancy.

Tested:
Function returns correct value based on the property.

Change-Id: I2b902972ef74ae5072560baac786c817f840059a
Signed-off-by: Matt Spinler <spinler@us.ibm.com>