RBMC: Set ReasonsForNoRedundancy prop on passive
If a BMC chooses to be passive due to some error case as determined by
Manager::determinePassiveRoleIfRequired(), then that means that
redundancy shouldn't be enabled because it can never be active.

The active BMC needs a way to know this, so the passive BMC will now add
an entry into its 'ReasonsForNoRedundancy' property which was previously
unused.  It doesn't need to put each individual reason into the
property, the fact that it isn't empty will be enough for the active BMC
to key off of.

Future commits will add support to the Sibling class for the active BMC
to read the property from the passive BMC and also then include this in
the checks to enable redundancy.

An alternative would have been to transfer every possible reason to the
active BMC and then have the active BMC explicitly check for it, but
that requires more work each time a new one is added and there will be
an error log created by the passive BMC with the actual reason anyway.

If necessary in the future code can be added to check if the problems
have been resolved so the property can be cleared.

Tested:

Set a failed status on the system inventory so the BMC is passive:
```
phosphor-rbmc-state-manager[8527]: Role = xyz.openbmc_project.State.BMC.Redundancy.Role.Passive due to: System inventory is not available
```

The property then has a value:
```
$ busctl get-property xyz.openbmc_project.State.BMC.Redundancy \
  /xyz/openbmc_project/state/bmc0 \
  xyz.openbmc_project.State.BMC.Redundancy \
  ReasonsForNoRedundancy
as 1 "xyz.openbmc_project.State.BMC.Redundancy.ReasonForNoRedundancy.SiblingCannotBeActive"
```

Change-Id: I93515ff2e04d5cf2f823a2ace7eb1a37a5e75618
Signed-off-by: Matt Spinler <spinler@us.ibm.com>