RBMC: Check if passive can be active or not
If the passive BMC has its own reason that redundancy can't be enabled,
it's because something is preventing it from ever being the active BMC
so redundancy should not be enabled.

Start including the check for that when determining if redundancy can be
enabled.  There is a new 'siblingCannotBeActive' member added to the
Input structure for checking redundancy which will be filled in with the
recent 'getHasReasonForNoRedundancy' sibling interface.  It will use the
new 'ReasonForNoRedundancy' value 'SiblingCannotBeActive'.

This also adds a line to rbmctool on the passive BMC to indicate if it
is OK to ever be active or not.

Now that this is being checked, the code could stop explicitly checking
if the sibling is provisioned or not when determining redundancy since
this new check would cover it. For now it will be left in since the
FFDC will be slightly better.

Tested:

```
phosphor-rbmc-state-manager[878]: Redundancy must be disabled
phosphor-rbmc-state-manager[878]: No redundancy because: SiblingCannotBeActive
```

rbmctool on the passive (new field at the bottom):
```
Local BMC
-----------------------------
Role:                 Passive
BMCPosition:          1
Redundancy Enabled:   false
BMC State:            Ready
Failovers Allowed:    false
Failover In Progress: false
FW Version Hash:      0204A8B8
Provisioned:          true
Role Reason:          System inventory is not available
Cannot Be Active:     true
```

Change-Id: I99851943e9144367ed3a0118c946d430c65a2d71
Signed-off-by: Matt Spinler <spinler@us.ibm.com>